#88: making the execute method more readable

+++ puppet/environments/box/modules/profiles/files/box/cli/scripts/Command/Share.php
@@ -7,8 +7,6 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use \Twig\Environment;
-use Symfony\Component\Process\Exception\ProcessFailedException;
-use Symfony\Component\Process\Process;
 
 class Share extends Command
 {
@@ -32,8 +30,6 @@ class Share extends Command
                 "Web server root",
                 '/var/www'
             );
-
-        //add optional run in background
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
@@ -45,33 +41,49 @@ class Share extends Command
         $this->www = $input->getOption('www');
         $this->site = $input->getArgument('site');
 
-        //going to need to port anything www or site related
+        $this->_check();
 
+        $this->_generateVhost();
 
-        //going to need to check (site exists)
-        $this->check($input, $output);
-        
-        $loader = new \Twig\Loader\FilesystemLoader('Command/templates/');
-        $this->twig = new \Twig\Environment($loader);
+        $this->_launchNgrok();
+    }
+
+    protected function _check()
+    {
+        if (!file_exists($this->www . "/" . $this->site)) {
+            throw new \RuntimeException(sprintf('Site not found: %s', $this->site));
+        }
 
         if (file_exists($this->vhost_dir . '/' . $this->vhost_file))
         {
             `sudo rm -f $this->vhost_dir/$this->vhost_file`;
             `sudo /etc/init.d/apache2 restart > /dev/null 2>&1`;
         }
+    }
+
+    protected function _generateVhost()
+    {
+        $loader = new \Twig\Loader\FilesystemLoader('Command/templates/');
+        $this->twig = new \Twig\Environment($loader);
 
         $template = $this->twig->load('ngrok_vhost.twig');
 
+        //now render the template with full variables inserted
         $vhost_config = $template->render(['path_to_site' => $this->www . "/" . $this->site, 'site' => $this->site]);
 
+        //create the writable tmp file
         file_put_contents($this->tmp_dir . "/" . $this->vhost_file, $vhost_config);
 
         `sudo tee /etc/apache2/sites-available/$this->vhost_file < $this->tmp_dir/$this->vhost_file`;
 
-        `sudo a2ensite $this->vhost_file`;
-        `sudo /etc/init.d/apache2 restart > /dev/null 2>&1`;
         `rm -f $this->tmp_dir/$this->vhost_file`;
 
+        `sudo a2ensite $this->vhost_file`;
+        `sudo /etc/init.d/apache2 restart > /dev/null 2>&1`;
+    }
+
+    protected function _launchNgrok()
+    {
         //so launch in the background
         `screen -d -m ngrok http $this->site.test:80`;
 
@@ -82,15 +94,7 @@ class Share extends Command
             $json = json_decode($result);
         } while (!isset($json->tunnels[0]->public_url));
 
-
         //then switch screens for the user
         `screen -r`;
-    }
-
-    protected function check(InputInterface $input, OutputInterface $output)
-    {
-        if (!file_exists($this->www . "/" . $this->site)) {
-            throw new \RuntimeException(sprintf('Site not found: %s', $this->site));
-        }
     }
 }
